Create fixes directory
The fixes/ directory will be recursively copied to /usr/lib/ganeti/fixes
on the bootstrap OS. It contains a few executable scripts which will be
run with run-parts. If these are to be python scripts, most of their
code should be contained in modules in fixes/fixlib/, so that they can
be unit tested.

This patch:
- Updates p2v-target.conf.in so that python will be available on the
  bootstap os
- Adds the hook script to install the fixes directory on the boostrap OS
- Creates the fixes/fixlib directory where python modules for the
  scripts will live

Signed-off-by: Ben Lipton <benlipton@google.com>
Signed-off-by: Iustin Pop <iustin@google.com>
[iustin@google.com: fixed copyright headers and whitespace]
Reviewed-by: Iustin Pop <iustin@google.com>