Better feedback for the user
p2v_transfer.py now shows the user a summary of what it is doing.
Additionally, there is now a check to see if a remote command is taking
longer than a minute. All it does is ask the user to try pinging the
machine and if it works to be patient, but it's better than no message
at all.

Signed-off-by: Ben Lipton <benlipton@google.com>
Signed-off-by: Iustin Pop <iustin@google.com>
Reviewed-by: Iustin Pop <iustin@google.com>