Better handling of source filesystems
Break out reading of source OS /etc/fstab into its own function, so we
only have to do it once. Using the results we can more reliably find the
amount of swap on the source machine. Also, we can mount and unmount all
of the disks that would normally be part of the source filesystem,
instead of being stuck with only the / partition.

Signed-off-by: Ben Lipton <benlipton@google.com>
Signed-off-by: Iustin Pop <iustin@google.com>
Reviewed-by: Iustin Pop <iustin@google.com>